Tidy of form.js
