+++ app/src/js/main.js
@@ -1,10 +1,10 @@
-import formSetup from './form';
+import formCtrl from './formCtrl';
 import fileUpload from './file-upload';
 import validation from './validation';
 import { checkSupport } from './v/index';
 
 checkSupport();
 
-formSetup.init();
+formCtrl.init();
 fileUpload.init();
 validation.init();
